+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New _go to definition_ feature, via `vscode.DefinitionProvider` class. Supports loop, functions, and regular assignment definitions.
+- Provide auto-completion for defined local variables.
 
 ### Fixed